Convert python connector from file to package
Created a quilt package for code clarity and compatibility with pip.
