+++ quilt/__init__.py
@@ -1,0 +1,18 @@
+
+from .util import File, Quilt
+from .table import Table, Column
+from .connection import Connection
+
+try:
+    import pandas
+    PANDAS = True
+except:
+    PANDAS = False
+
+try:
+    import psycopg2
+    import sqlalchemy
+    SQLALCHEMY = True
+except:
+    SQLALCHEMY = False
+
